+++ src/external/ours/library/archive/src/shared/AutoDeltaPackedMap.h
@@ -99,8 +99,8 @@ namespace Archive
 		char temp[200];
 		typename AutoDeltaMap<KeyType, ValueType, ObjectType>::Command c;
 
-		Archive::put(target, countCharacter(buffer,':'));
-		Archive::put(target, static_cast<size_t>(0)); // baselineCommandCount
+		Archive::put(target, static_cast<unsigned int>(countCharacter(buffer,':')));
+		Archive::put(target, static_cast<unsigned int>(0)); // baselineCommandCount
 		
 		int tempPos = 0;
 		for (std::string::const_iterator i=buffer.begin(); i!=buffer.end(); ++i)
@@ -127,8 +127,8 @@ namespace Archive
 		char temp[200];
 		
 		typename AutoDeltaMap<KeyType, ValueType, ObjectType>::Command c;
-		size_t commandCount;
-		size_t baselineCommandCount;
+		unsigned int commandCount;
+		unsigned int baselineCommandCount;
 
 		Archive::get(source, commandCount);
 		Archive::get(source, baselineCommandCount);
